various updates
