+++ .gitattributes
@@ -9,9 +9,9 @@
 /.github/pull_request_template.md linguist-generated
 /.github/workflows/auto-approve.yml linguist-generated
 /.github/workflows/build.yml linguist-generated
-/.github/workflows/dependabot-security-alerts.yml linguist-generated
 /.github/workflows/pull-request-lint.yml linguist-generated
 /.github/workflows/release.yml linguist-generated
+/.github/workflows/security.yml linguist-generated
 /.github/workflows/triage.yml linguist-generated
 /.github/workflows/upgrade-main.yml linguist-generated
 /.gitignore linguist-generated
