+++ .github/workflows/security.yml
@@ -1,19 +1,18 @@
 # ~~ Generated by projen. To modify, edit .projenrc.ts and run "npx projen".
 
-name: dependabot-security-alerts
+name: security
 on:
   schedule:
-    - cron: 27 * * * *
+    - cron: 0 0 * * *
   workflow_dispatch: {}
 jobs:
-  dependabot_security_alert:
-    name: dependabot-security-alert
+  scan:
     runs-on: ubuntu-latest
     permissions:
       security-events: read
       issues: write
     steps:
-      - name: Run Script
+      - name: scan
         uses: cdk8s-team/cdk8s-dependabot-security-alerts@main
         env:
           GITHUB_TOKEN: ${{ secrets.PROJEN_GITHUB_TOKEN }}
