+++ .gitignore
@@ -60,5 +60,5 @@ tsconfig.json
 !/.github/ISSUE_TEMPLATE/bug.md
 !/.github/ISSUE_TEMPLATE/feature-request.md
 !/SECURITY.md
+!/.github/workflows/security.yml
 !/.github/workflows/triage.yml
-!/.github/workflows/dependabot-security-alerts.yml
