+++ .projen/files.json
@@ -8,9 +8,9 @@
     ".github/pull_request_template.md",
     ".github/workflows/auto-approve.yml",
     ".github/workflows/build.yml",
-    ".github/workflows/dependabot-security-alerts.yml",
     ".github/workflows/pull-request-lint.yml",
     ".github/workflows/release.yml",
+    ".github/workflows/security.yml",
     ".github/workflows/triage.yml",
     ".github/workflows/upgrade-main.yml",
     ".gitignore",
